esdfs: add package-derived permission support
This commit implements the new derived permissions functionality in
the "sdcard" service that was introduced in Android 4.4.  The file
system is still case-sensitive and case-preserving and uses fixed
permissions for the majority of the tree, but the "Android" tree is
handled specially.  User space feeds the Android package list to the
driver, which is used to determine the ownership of package-specific
directories under "Android/data".

Change-Id: Ic7038c4eafbda36d0c527ea45e668d15c54cab65
CRs-fixed: (CR)
Signed-off-by: Russ Knize <Russ.Knize@motorola.com>