icnss: Implement Wifi Detune
Vote for the Wifi Detune when Wifi is Enabled
and remove the vote when Wifi is disabled via
the icnss driver. The Wifi Detune controls the
state of the antenna tuners when Wifi is active
and WAN is disabled.

Change-Id: If77e128fe49c970b6489a31a9f19614141abf76f
Signed-off-by: Ravi Vembu <raviv@motorola.com>
Reviewed-on: https://gerrit.mot.com/957437
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>