arm/dts:Add property single path usbin switch
For beckham, HW use ncp3902 instead of ncp3901, it is a
single path usbin switch which just control usbin path. And
as a result, its SW logic will be different with ncp3901. So
we add this property for driver to distinguish the SW logic
for them. This commit is for sdm636.

Change-Id: Ie13e4f5c7ae40a8081a40483450f9fed3addd6eb
Signed-off-by: weiweij <weiweij@motorola.com>
Reviewed-on: https://gerrit.mot.com/1093032
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key